# docs/API.yaml
openapi: 3.0.3
info:
  title: Prosto i Zdorovo API
  version: 0.1.0
  description: Backend stack uses Spring Boot + Kotlin.
servers:
  - url: http://localhost:3000
paths:
  /api/webhook/json-rpc/:
    post:
      summary: JSON-RPC 2.0 endpoint used by frontend
      description: |
        Base URL used by frontend RPC client (see front/handmade). Request is sent to
        /api/webhook/json-rpc/?method=<method> with JSON-RPC 2.0 body.
        Known methods (MVP):
          - get-appointments
          - get-appointment-types
          - specialist-available
          - get-goals
          - free-slots
          - create-appointment
          - confirm-appointment
          - search-form
          - get-form-template
          - form-submission
          - get-recommendations
          - recommendations-specialist
          - get-profile
          - get-chats
          - get-messages
      parameters:
        - in: query
          name: method
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RpcRequest"
      responses:
        "200":
          description: RPC response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RpcResponse"
...
components:
  schemas:
    RpcRequest:
      type: object
      properties:
        jsonrpc:
          type: string
          example: "2.0"
        method:
          type: string
          description: RPC method name
        params:
          type: object
          description: Method params (see specific method schemas)
        id:
          type: string
        skipCache:
          type: boolean
      required: [jsonrpc, method]
    RpcResponse:
      type: object
      properties:
        jsonrpc:
          type: string
          example: "2.0"
        id:
          type: string
        result:
          type: object
          description: Method result (see specific method schemas)
        error:
          $ref: "#/components/schemas/RpcError"
    RpcError:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
...